--- cj-connector/src/main/kotlin/com/alexbogovich/cj/Errors.kt
@@ -1,3 +1,0 @@
-package com.alexbogovich.cj
-
-class CjRequestException(msg: String) : RuntimeException(msg)
