+++ cj-connector/src/main/kotlin/net/besttoolbars/cj/Errors.kt
@@ -1,0 +1,3 @@
+package net.besttoolbars.cj
+
+class CjRequestException(msg: String) : RuntimeException(msg)
