--- impact-radius-connector/src/main/kotlin/com/alexbogovich/impactradius/Errors.kt
@@ -1,4 +1,0 @@
-package com.alexbogovich.impactradius
-
-data class ImpactRadiusRequestException(val statusCode: Int, val body: String) :
-    RuntimeException("Bad response with code $statusCode. Body: $body")
