--- impact-radius-connector/src/main/kotlin/com/alexbogovich/impactradius/Query.kt
@@ -1,8 +1,0 @@
-package com.alexbogovich.impactradius
-
-//data class ImpactRadiusGenericQuery(
-//    val accountSid: String,
-//    val token: String,
-//    val page: Int = 1,
-//    val pageSize: Int = 100
-//)
