+++ impact-radius-connector/src/main/kotlin/net/besttoolbars/impactradius/Errors.kt
@@ -1,0 +1,4 @@
+package net.besttoolbars.impactradius
+
+data class ImpactRadiusRequestException(val statusCode: Int, val body: String) :
+    RuntimeException("Bad response with code $statusCode. Body: $body")
