+++ impact-radius-connector/src/main/kotlin/net/besttoolbars/impactradius/Query.kt
@@ -1,0 +1,8 @@
+package net.besttoolbars.impactradius
+
+//data class ImpactRadiusGenericQuery(
+//    val accountSid: String,
+//    val token: String,
+//    val page: Int = 1,
+//    val pageSize: Int = 100
+//)
