--- rakuten-connector/src/main/kotlin/com/alexbogovich/rakuten/Errors.kt
@@ -1,4 +1,0 @@
-package com.alexbogovich.rakuten
-
-data class RakutenRequestException(val statusCode: Int, val body: String) :
-    RuntimeException("Bad response with code $statusCode. Body: $body")
