+++ rakuten-connector/src/main/kotlin/net/besttoolbars/rakuten/Errors.kt
@@ -1,0 +1,4 @@
+package net.besttoolbars.rakuten
+
+data class RakutenRequestException(val statusCode: Int, val body: String) :
+    RuntimeException("Bad response with code $statusCode. Body: $body")
