+++ shared-xml-mapper/src/main/kotlin/net/besttoolbars/connectors/shared/FloatDeserializer.kt
@@ -1,0 +1,13 @@
+package net.besttoolbars.connectors.shared
+
+import com.fasterxml.jackson.core.JsonParser
+import com.fasterxml.jackson.databind.DeserializationContext
+import com.fasterxml.jackson.databind.JsonDeserializer
+import com.fasterxml.jackson.databind.deser.std.StringDeserializer
+
+object FloatDeserializer : JsonDeserializer<Float?>() {
+    override fun deserialize(p: JsonParser?, ctxt: DeserializationContext?): Float? {
+        val result: String? = StringDeserializer.instance.deserialize(p, ctxt)
+        return result?.toFloatOrNull()
+    }
+}
